Improved dependency aviability check
Moved the bootstrap code to an appropriate files
Checking if the packages are installed without importing the modules.
Also add a check on animate object because auto-install doesn't call register/unregister.
Added uninstallation of packages on unregistration.
Added Exception on some errors so they are displayer as a Blender popup instead on only a print on console.
Increased version.